refactor: avoid events for deleting
